Tests/LibWeb/Text/expected/css/invalid-selector-in-has.txt

:has(:yakthonk) should be invalid: PASS
